SQLServer updates, data type one-to-many encoding, codelist insert quote
* SQL Server updates: using bigint instead of int as primary key data
type, bounding box for spatial index only when configured in map entry
(no longer a default value) 
* Two conversion rules have been added to encode one-to-many
relationships with data types. 
* Single quote in values of string expressions of an insert statement
for a code list table are now escaped (replaced by two single quotes).
* Name of PODS specific fields that can be added to codelist tables are
now configurable.

+++ src/test/resources/reference/sql/defaultValues/ddl/sqlserver/F_types/Test6_Initial_Values.sql
@@ -1,6 +1,6 @@
 CREATE TABLE T6_Class1 (
 
-   _id int NOT NULL PRIMARY KEY,
+   _id bigint NOT NULL PRIMARY KEY,
    propBooleanFalse bit NOT NULL DEFAULT 0,
    propBooleanTrue bit NOT NULL DEFAULT 1,
    propInteger int NOT NULL DEFAULT 3,
@@ -10,7 +10,7 @@ CREATE TABLE T6_Class1 (
 
 CREATE TABLE T6_Class1_propStrings (
 
-   T6_Class1_id int NOT NULL,
+   T6_Class1_id bigint NOT NULL,
    propStrings nvarchar(max) NOT NULL DEFAULT 'abc',
    PRIMARY KEY (T6_Class1_id, propStrings)
 );
